Remove test-render-templates from makefile
test-render-templates called make from a bash script. The nested
invocation caused an error on centos. When 'make targets' is called
from within 'test-targets.sh' the phrase 'make[1]: Leaving directory...'
would be included in the array of targets. This could be fixed by using
the flag `--no-print-directory` but that's not reasonable to ask all users
to do.